Split monomorphized type specialization
